6_deployment/fruit-app/requirements.txt
absl-py==0.6.1
alabaster==0.7.12
astor==0.7.1
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.15.4
numpydoc==0.8.0
olefile==0.46
packaging==18.0
Pillow==4.2.1
protobuf==3.6.1
Pygments==2.3.1
pyparsing==2.3.0
pytz==2018.7
PyYAML==4.2b1
requests==2.21.0
scipy==1.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.2
Werkzeug==0.14.1
WTForms==2.2.1
